Add ExceptionHandler and enhance TaskController's update
A global exception handler for runtime exceptions is introduced called 'ExceptionHandlerController', specifically for HttpMessageNotReadableException. An update() function in TaskController is modified to return bad request if task is not found, and unauthorised if logged in user is not the owner. The method name in Utils is also edited from 'getNonNullProperties' to 'copyNonNullProperties' for clarity.